feat: add apikey authorization
Adds a solution for API Key generation and authorization enforcement.

This is done with basic JWT tokens that can be manually generated with a script for different envs. This makes it relatively simple for us to rotate the secret and invalidate all existing "keys" if neccesary. A NestJS guard is used to apply authorization logic to different controllers or specific endpoints. Swagger has also been configured to allow easy "logins" for local testing.

To track request origins, a uuid is embedded in all apiKeys. This apikey id is now included in all request logs under the property `apiKeyId`.

A helper script is included for generating new keys for any env. See documentation update for details.